+++ src/app/products/components/product/product.component.ts
@@ -1,6 +1,11 @@
-import { Component, Input, Output, EventEmitter, ChangeDetectionStrategy } from '@angular/core';
+import { Component, Input, Output, EventEmitter, ChangeDetectionStrategy, OnInit, OnDestroy } from '@angular/core';
+import { ActivatedRoute, Router } from '@angular/router';
+
+import { Subscription } from 'rxjs';
 
 import { ProductModel } from '../../../shared/models/product';
+import { CartService } from '../../../cart/services/cart.service';
+import { ProductsService } from '../../services/products-service';
 
 @Component({
   selector: 'app-product',
@@ -8,14 +13,42 @@ import { ProductModel } from '../../../shared/models/product';
   styleUrls: ['./product.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class ProductComponent {
+export class ProductComponent implements OnInit, OnDestroy {
   @Input() product: ProductModel;
-  @Output() productBought = new EventEmitter<void>();
+  @Output() productSelected = new EventEmitter<number>();
 
-  constructor() { }
+  private sub: Subscription;
+
+  constructor(
+    private router: Router,
+    private route: ActivatedRoute,
+    private productsService: ProductsService,
+    private cartService: CartService
+  ) { }
+
+  ngOnInit(): void {
+    const id = +this.route.snapshot.paramMap.get('productID');
+    if (id > 0) {
+      const observer = {
+        next: (product: ProductModel) => {
+          this.product = { ...product };
+        },
+        error: (err: any) => console.log(err)
+      };
+      this.sub = this.productsService.getProduct(id).subscribe(observer);
+    }
+  }
+
+  ngOnDestroy(): void {
+    this.sub?.unsubscribe();
+  }
 
   onBuy(): void {
     console.log(`Congratulations! You bought: ${this.product.name} by ${this.product.price}!`);
-    this.productBought.emit();
+    this.cartService.addProduct(this.product);
+  }
+
+  onGoToProduct(): void {
+    this.router.navigate([`/product/${this.product.id}`]);
   }
 }